+++ src/onegov.activity/onegov/activity/matching/utils.py
@@ -1,16 +1,23 @@
+from datetime import timedelta
+
 from onegov.activity import log
 from onegov.activity.utils import dates_overlap
 from sortedcontainers import SortedSet
 
 
-def overlaps(booking, other):
+def overlaps(booking, other, minutes_between=0):
     """ Returns true if the two given bookings overlap. """
 
     assert booking != other
 
+    offset = timedelta(seconds=minutes_between / 2 * 60)
+
+    # make sure that 11:00 - 12:00 and 12:00 - 13:00 are not overlapping
+    ms = timedelta(microseconds=1)
+
     return dates_overlap(
-        tuple((b.start, b.end) for b in booking.dates),
-        tuple((o.start, o.end) for o in other.dates),
+        tuple((b.start - offset, b.end + offset - ms) for b in booking.dates),
+        tuple((o.start - offset, o.end + offset - ms) for o in other.dates),
     )